added risk score calculator and did some refactorization
