+++ src/db/models/installation.ts
