+++ src/types/RiskScore.ts
@@ -1,0 +1,5 @@
+type RiskScore = {
+  score: number;
+};
+
+export default RiskScore;
